// scripts/models/AddressDetail.ts
export class AddressDetail {
  name: string | null = ""
  email: string | null = ""
  phone: string | null = ""
  addressLine1: string | null = ""
  addressLine2: string | null = ""
  addressLine3: string | null = ""
  addressPostcode: string | null = ""
  addressCity: string | null = ""
  addressState: string | null = ""
  addressCountry: string | null = ""
}